groupByLangをユーティリティとして分離
